+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Refactored `getBasicBoxScoreStats` and `GetMatchups` to use `networkHeaders` and `DocumentRetriever`
+- Logging is more concise for `getBasicBoxScoreStats`
+- Slight refactor of `getBasicBoxScoreStats` to use shared components
 ### Added
-- `NBAAdvBoxScoreStats` data model for basketball-reference NBA Advanced box score stats
+- `NBAAdvBoxScoreStats` data model for basketball-reference.com NBA Advanced box score stats
+- Created shared.go for shared components for scraping NBA box score data from basketball-reference.com
+- Created functions to extract player ID and transform player minutes played
 
 ## [0.1.0-alpha] - 2025-02-25
 ### Added